Rename Consumer namespace to Contained.
