+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/Contained.meta
