+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/Contained/IUnityContained.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 975ee4091d1752d4e8ae566044218a88
