+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/Contained/IUnityDependentExtensions.cs
@@ -3,7 +3,7 @@ using Chopsticks.Dependencies.Resolutions;
 using System;
 using UnityEngine;
 
-namespace Chopsticks.Dependencies.Consumers
+namespace Chopsticks.Dependencies.Contained
 {
     /// <summary>
     /// Provides extensions to Unity dependents.
