+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/Contained/IUnityDependentExtensions.cs.meta
